+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * New rules:
 
+  - `any_is_na` now also reports `NA %notin% x` cases (#470, @Yousa-Mirage)
   - `notin` (#459, @Yousa-Mirage)
 
 * Jarl is now available on PyPI under the name `jarl-linter`, enabling its